对接前端接口调整

+++ app/admin/internal/logic/admin_info/admin_info.go
@@ -87,5 +87,5 @@ func Login(ctx context.Context, name, password string) (token string, expire tim
 	}
 
 	// 5. 生成JWT Token
-	return utility.GenerateToken(admin.Id)
+	return utility.GenerateToken(uint32(admin.Id))
 }
